% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/add_zero.R
\name{add_zero}
\alias{add_zero}
\title{Add zero to front of x and make x a character string}
\usage{
add_zero(x, desired_length)
}
\arguments{
\item{x}{object for whome the "0" must be added in the beginning. Can be any object what can be converted to char.}

\item{desired_length}{integer of what length should the output be.}
}
\value{
character string
}
\description{
Function for adding "0" to front of character dates for example.
}
