% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_age.R
\name{calc_age}
\alias{calc_age}
\title{Function for calculating age at eventdate}
\usage{
calc_age(eventdate, birthdate, accuracy = "years")
}
\arguments{
\item{eventdate}{lubridate date when the age is needed}

\item{birthdate}{lubridate date of person's birthdate}
}
\value{

}
\description{
The function helps calculating the age of person at the eventdate
}
\seealso{
Other date functions: 
\code{\link{filter_date_first}()},
\code{\link{filter_date_last}()},
\code{\link{filter_date_nearest}()}
}
\concept{date functions}
