% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{classify_charlson}
\alias{classify_charlson}
\title{Extends the original data with the classification table and keeps data in long format.
Computes classification table and attaches it to original data. Data stays in long format.}
\usage{
classify_charlson(.data, icd_codes)
}
\arguments{
\item{.data}{tibble of register data which we want to study}

\item{icd_codes}{name of the variable holding ICD-codes}
}
\value{
data frame
}
\description{
Extends the original data with the classification table and keeps data in long format.
Computes classification table and attaches it to original data. Data stays in long format.
}
\examples{
\dontrun{
}

}
