% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify_data_long.R
\name{classify_data_long}
\alias{classify_data_long}
\title{Extends the original data with the classification table and keeps data in long format.}
\usage{
classify_data_long(.data, id, icdcodes, diag_tbl, fill = 0)
}
\arguments{
\item{.data}{tibble of register data which we want to study}

\item{icdcodes}{name of the variable holding ICD-codes}

\item{diag_tbl}{tibble which holds the classification details: needs to have variables 'regex' and 'label'
'regex' must hold a string with a regular expression defining classes.
'regex.rm' is optional, defines exceptions to 'regex' (these are not in the group they are named in)
'label' defines the names of the variables of classes (e.g. comorbidity indicators)}
}
\value{

}
\description{
Computes classification table and attaches it to original data. Data stays in long format.
}
\examples{
\dontrun{
# we calculate the table which can be used for classification.
# 'ostprekoh' is the register data containing 'lomno1' individual id number and data 'dg' contains the register data of disease diagnoses.

dat<-ostprekoh \%>\%
left_join(dg \%>\% select(lomno1,KOODI1,tulopvm,lahtopvm,icd),by="lomno1") \%>\%
  mutate(Postituspvm=ymd(Postituspvm)) \%>\%
  datefilter(indexdate=Postituspvm,range=years(2),tulopvm,lahtopvm) \%>\%
  classify_data_long(icdcodes=KOODI1,diag_tbl=sel_classes) \%>\%
  filter(match>0)
}

}
