% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regstudies_24032020.R
\name{classify_no_rm}
\alias{classify_no_rm}
\title{Deprecated: Classify diagnosis codes (no exceptions)}
\usage{
classify_no_rm(.data, icdcodes, diag_tbl, return_binary = TRUE)
}
\arguments{
\item{.data}{tibble of register data which we want to study}

\item{icdcodes}{name of the variable holding ICD-codes}

\item{diag_tbl}{tibble which holds the classification details: needs to have variables 'regex' and 'label'
'regex' must hold a string with a regular expression defining classes.
'regex.rm' is optional, defines exceptions to 'regex' (these are not in the group they are named in)
'label' defines the names of the variables of classes (e.g. comorbidity indicators)}

\item{return_binary}{returns 0 or 1 if set as TRUE, TRUE/FALSE if set up as FALSE}
}
\value{

}
\description{
Computes classification table which can be attached to original data using left_join(). Does not utilise exceptions to class definitions via 'regex.rm'
}
\examples{
# we calculate the table which can be used for classification.
# 'ostprekoh' is the register data containing 'lomno1' individual id number and data 'dg' contains the register data of disease diagnoses.
ostprekoh \%>\%
  left_join(dg \%>\% select(lomno1,KOODI1,tulopvm,lahtopvm),by="lomno1") \%>\%
  mutate(Postituspvm=ymd(Postituspvm)) \%>\%
  datefilter(indexdate=Postituspvm,range=years(2),tulopvm,lahtopvm) \%>\% # filtering the diagnosis codes which are in the interval for each individual!
  classify_long(icdcodes=KOODI1,diag_tbl=sel_classes)

}
