% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regstudies_24032020.R
\name{classifydata}
\alias{classifydata}
\title{Deprecated: Extends the original data with the classification table and makes wide data.}
\usage{
classifydata(.data, id, icdcodes, diag_tbl, return_binary = TRUE)
}
\arguments{
\item{.data}{tibble of register data which we want to study}

\item{id}{personal identification number}

\item{icdcodes}{name of the variable holding ICD-codes}

\item{diag_tbl}{tibble which holds the classification details: needs to have variables 'regex' and 'label'
'regex' must hold a string with a regular expression defining classes.
'regex.rm' is optional, defines exceptions to 'regex' (these are not in the group they are named in)
'label' defines the names of the variables of classes (e.g. comorbidity indicators)}

\item{return_binary}{returns 0 or 1 if set as TRUE, TRUE/FALSE if set up as FALSE}
}
\value{

}
\description{
Computes classification table and attaches it to original data. Attached data is wide format (many new variables).
}
\examples{
# we calculate the table which can be used for classification.
# 'ostprekoh' is the register data containing 'lomno1' individual id number and data 'dg' contains the register data of disease diagnoses.

}
