% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_date.R
\name{filter_date}
\alias{filter_date}
\title{Filtering of datetime data within interval.}
\usage{
filter_date(.data, indexdate, range = years(2), ...)
}
\arguments{
\item{indexdate}{index date (which date variable is to be compared with register data)}

\item{range}{the distance from index date (lubridate format, e.g years(1), weeks(10), days(20) etc.)}

\item{...}{datetime variables which need to be within the range. May be any number (>1) of variables.  If one variable is within interval, then that row is included in the data (Applies OR operator).}
}
\value{
returns a tibble object with filtered rows
}
\description{
Filtering of datetime data within interval.
}
\examples{
\dontrun{
'ostprekoh' is the register data containing 'lomno1' individual id number and data 'dg' contains the register data of disease diagnoses.

dat <- ostprekoh \%>\%
left_join(dg \%>\% select(lomno1,KOODI1,tulopvm,lahtopvm,icd),by="lomno1") \%>\%
  mutate(Postituspvm=ymd(Postituspvm)) \%>\%
  filter_date(indexdate=Postituspvm,range=years(2),tulopvm,lahtopvm)
}
}
