% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_date_hosp.R
\name{filter_date_hosp}
\alias{filter_date_hosp}
\title{Filtering of hospitalisation interval data within study interval.}
\usage{
filter_date_hosp(
  .data,
  indexdate,
  time_before = years(2),
  time_after = days(0),
  admission_date,
  discharge_date
)
}
\arguments{
\item{indexdate}{index date (which date variable is to be compared with register data)}

\item{time_before}{the time before the index date what defines the start of filtering interval (lubridate format, e.g years(1), weeks(10), days(20) etc.)}

\item{time_after}{the time before the index date (default is 0 days) what defines the start of filtering interval (lubridate format, e.g years(1), weeks(10), days(20) etc.)}

\item{admission_date}{date of admission as lubridate format}

\item{discharge_date}{date of discharge as lubridate format}
}
\value{
returns a tibble object with filtered rows
}
\description{
Filtering of hospitalisation interval data within study interval.
}
\examples{
\dontrun{
#'ostprekoh' is the register data containing 'lomno1' individual id number and data 'dg' contains the register data of disease diagnoses.

dat <- cohort \%>\%
left_join(reg \%>\% select(personid,CODE1,admissiondate,dischargedate,icd),by="personid") \%>\%
  filter_date_hosp(indexdate=postingdate,
                   time_before=years(2),time_after=years(2),
                   admission_date=admissiondate,
                   discharge_date=dischargedate) \%>\%
  select(-study_interval,-hosp_interval)
head(dat)
}
}
