% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_date_in_ival}
\alias{filter_date_in_ival}
\title{Filtering of datetime data within a time interval.
The function finds rows for which \code{date_var} is within interval created based on \code{index_date}, \code{time_before} and \code{time_after}.}
\usage{
filter_date_in_ival(
  .data,
  date_var,
  index_date,
  time_before = years(2),
  time_after = days(0)
)
}
\arguments{
\item{date_var}{}

\item{index_date}{date or other lubridate variable as reference point around which to search}

\item{time_before}{the time before the index date what defines the start of filtering interval (lubridate format, e.g years(1), weeks(10), days(20) etc.)}

\item{time_after}{the time after the index date (default is 0 days) what defines the start of filtering interval (lubridate format, e.g years(1), weeks(10), days(20) etc.)}
}
\value{
returns a tibble object with filtered rows
}
\description{
Filtering of datetime data within a time interval.
The function finds rows for which \code{date_var} is within interval created based on \code{index_date}, \code{time_before} and \code{time_after}.
}
\examples{
\dontrun{
TODO:
x <- 1
}
}
\seealso{
\code{\link{filter_ival_olap_ival}} for date filtering by different intervals before and after the occurence

Other filter functions: 
\code{\link{filter_ival_olap_ival}()}
}
\concept{filter functions}
