% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_date_interval.R
\name{filter_date_interval}
\alias{filter_date_interval}
\title{Filtering of date interval data within study interval.
Can be used for filtering e.g. for hospital visits, which are date intervals (time periods).}
\usage{
filter_date_interval(
  .data,
  indexdate,
  time_before = years(2),
  time_after = days(0),
  admission_date,
  discharge_date
)
}
\arguments{
\item{indexdate}{index date (which date variable is to be compared with register data)}

\item{time_before}{the time before the index date what defines the start of filtering interval (lubridate format, e.g years(1), weeks(10), days(20) etc.)}

\item{time_after}{the time before the index date (default is 0 days) what defines the start of filtering interval (lubridate format, e.g years(1), weeks(10), days(20) etc.)}

\item{admission_date}{date of admission as lubridate format}

\item{discharge_date}{date of discharge as lubridate format}
}
\value{
returns a tibble object with filtered rows
}
\description{
Filtering of date interval data within study interval.
Can be used for filtering e.g. for hospital visits, which are date intervals (time periods).
}
\examples{
\dontrun{
d <- dplyr::left_join(cohort_data,reg_data)
filtered_d <- d \%>\% 
  filter_date_interval(indexdate = postingdate,
                       time_before=years(2),
                       admission_date=adm_date,
                       discharge_date=disc_date)
head(filtered_d)
}
}
\seealso{
\code{\link{filter_date_range}} for date filtering by fixed interval

Other filter functions: 
\code{\link{filter_date_range}()}
}
\concept{filter functions}
