% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_date_nearest}
\alias{filter_date_nearest}
\title{Get nearest value based on datetime variable}
\usage{
filter_date_nearest(.data, date_var, index_date, only_previous = FALSE)
}
\arguments{
\item{index_date}{datetime or other comparable variable as reference point around which to search}

\item{only_previous}{optional parameter (default=FALSE): should we search only previously observed data prior index_date?}

\item{.date}{data which contains datetime or other comparable variable}

\item{var}{datetime or other comparable variable}
}
\value{
The row which is the nearest .date
}
\description{
This function helps user to find data rows which have datetime value of interest \code{date_var} closest to specified \code{index_date} value.
User can use argument \code{only_previous} to handle if nearest row can be only before or also after the \code{index_date} date.
}
\examples{
\dontrun{
# searching for the first event of the each class (for each individual):
d \%>\%
  group_by(personid) \%>\%
  filter_date_nearest(adm_date,index_date=postingdate)
}

}
\seealso{
Other date functions: 
\code{\link{calc_age}()},
\code{\link{filter_date_first}()},
\code{\link{filter_date_last}()}
}
\concept{date functions}
