% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter_date_range.R
\name{filter_date_range}
\alias{filter_date_range}
\title{Filtering of datetime data within a time range.}
\usage{
filter_date_range(.data, indexdate, range = years(2), ...)
}
\arguments{
\item{indexdate}{index date (which date variable is to be compared with register data)}

\item{range}{the distance from index date (lubridate format, e.g years(1), weeks(10), days(20) etc.)}

\item{...}{datetime variables which need to be within the range. May be any number (>1) of variables.  If one variable is within interval, then that row is included in the data (Applies OR operator).}
}
\value{
returns a tibble object with filtered rows
}
\description{
Filtering of datetime data within a time range.
}
\examples{
\dontrun{
TODO:
x <- 1
}
}
\seealso{
\code{\link{filter_date_interval}} for date filtering by different intervals before and after the occurence

Other filter functions: 
\code{\link{filter_date_interval}()}
}
\concept{filter functions}
