% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_random_hetu.R
\name{generate_random_hetu}
\alias{generate_random_hetu}
\title{Function for generating random finnish personal identification number (Henkilotunnus)}
\usage{
generate_random_hetu(birthdate)
}
\arguments{
\item{birthdate}{this must be valid lubridate date between start of 1800 and end of 2000 centuries.}
}
\value{
character string henkilotunnus
}
\description{
The function is able to generate a nearly valid random finnish personal identification number.
}
