% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_age.R
\name{get_age}
\alias{get_age}
\title{Function for calculating age at eventdate}
\usage{
get_age(eventdate, birthdate, accuracy = "years")
}
\arguments{
\item{eventdate}{lubridate date when the age is needed}

\item{birthdate}{lubridate date of person's birthdate}
}
\value{

}
\description{
The function helps calculating the age of person at the eventdate
}
