% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_birthdate.R
\name{get_birthdate}
\alias{get_birthdate}
\title{Function for extracting birthdate from finnish personal identification number (Henkilotunnus)}
\usage{
get_birthdate(hetu)
}
\arguments{
\item{hetu}{valid personal identification number (henkilotunnus)}
}
\value{
lubridate::date indicating the birthdate
}
\description{
The function is able to extract birthdate of a person from a finnish personal identification number.
}
