% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_first.R
\name{get_first}
\alias{get_first}
\title{Get first value based on datetime variable}
\usage{
get_first(.date, var)
}
\arguments{
\item{.date}{data which contains datetime or other comparable variable}

\item{var}{datetime or other comparable variable}
}
\value{
The row which is the first (by which.min) in the .date
}
\description{
Get first value based on datetime variable
}
\examples{
\dontrun{
# searching for the first event of the each class (for each individual):
dat \%>\%
  filter(!is.na(label)) \%>\%
  group_by(lomno1,label) \%>\%
  get_first(tulopvm)
}
}
\seealso{
Other date functions: 
\code{\link{get_age}()},
\code{\link{get_last}()},
\code{\link{get_nearest}()}
}
\concept{date functions}
