% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_gender.R
\name{get_gender}
\alias{get_gender}
\title{Function for extracting gender from finnish personal identification number (Henkilotunnus)}
\usage{
get_gender(hetu)
}
\arguments{
\item{hetu}{valid personal identification number (henkilotunnus)}
}
\value{
factor object indicating gender of person
}
\description{
The function is able to extract gender of a person from a finnish personal identification number.
}
