% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_last.R
\name{get_last}
\alias{get_last}
\title{Get last value based on datetime variable}
\usage{
get_last(.data, var)
}
\arguments{
\item{.data}{data which contains datetime or other comparable variable}

\item{var}{datetime or other comparable variable}
}
\value{
The row which is the first (by which.max) in the .date
}
\description{
Get last value based on datetime variable
}
\examples{

\dontrun{
# searching for the first event of the each class (for each individual):
dat \%>\%
  filter(!is.na(label)) \%>\%
  group_by(personid,label) \%>\%
  get_last(adm_date)
}
}
\seealso{
Other date functions: 
\code{\link{get_age}()},
\code{\link{get_first}()},
\code{\link{get_nearest}()}
}
\concept{date functions}
