% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_na_vars.R
\name{get_na_vars}
\alias{get_na_vars}
\title{Function for extracting names of variables which have any NA values}
\usage{
get_na_vars(.data)
}
\arguments{
\item{.data}{data to be studied}
}
\value{
returns a two-column tibble holding variable name as first column ('variable') and 'contains_na' as second column.
}
\description{
Function for extracting names of variables which have any NA values
}
