% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regstudies_24032020.R
\name{getfirst}
\alias{getfirst}
\title{Get first value based on datetime variable}
\usage{
getfirst(.date, var)
}
\arguments{
\item{.date}{data which contains datetime or other comparable variable}

\item{var}{datetime or other comparable variable}
}
\value{
The row which is the first (by which.min) in the .date
}
\description{
Get first value based on datetime variable
}
\examples{
# searching for the first event of the each class (for each individual):
dat \%>\%
  filter(!is.na(label)) \%>\%
  group_by(lomno1,label) \%>\%
  getfirst(tulopvm)

}
