% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regstudies_24032020.R
\name{getnearest}
\alias{getnearest}
\title{Get nearest value based on datetime variable}
\usage{
getnearest(.date, var, centroid, only_previous = FALSE)
}
\arguments{
\item{.date}{data which contains datetime or other comparable variable}

\item{var}{datetime or other comparable variable}

\item{centroid}{datetime or other comparable variable as reference point around which to search}

\item{only_previous}{optional parameter (default=FALSE): should we search only previously observed data prior centroid?}
}
\value{
The row which is the nearest .date
}
\description{
Get nearest value based on datetime variable
}
\examples{
# searching for the first event of the each class (for each individual):
dat \%>\%
  getnearest(tulopvm,centroid=Postituspvm)

}
