% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join0.R
\name{left_join0}
\alias{left_join0}
\title{left_join with fill option}
\usage{
left_join0(x, y, fill = 0)
}
\arguments{
\item{x}{tibble data to used as left in a join}

\item{y}{tibble data to used as right in a join}

\item{fill}{value to be filled into NA values}
}
\value{
resulting tibble after join and fill
}
\description{
Allows user to do left_join and fill missing NA values with 0 or other optional value
}
