% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classify.R
\name{make_classify_table}
\alias{make_classify_table}
\title{Classify diagnosis codes to long format with exceptions}
\usage{
make_classify_table(.data, icdcodes, diag_tbl, return_binary = TRUE)
}
\arguments{
\item{.data}{tibble of register data which we want to study}

\item{icdcodes}{name of the variable holding ICD-codes}

\item{diag_tbl}{tibble which holds the classification details: needs to have variables 'regex' and 'label'
'regex' must hold a string with a regular expression defining classes.
'regex.rm' is optional, defines exceptions to 'regex' (these are not in the group they are named in)
'label' defines the names of the variables of classes (e.g. comorbidity indicators)}
}
\value{
Returns a tibble containing classification table which can be joined to initial data
}
\description{
Computes classification table which can be attached to original data using left_join().
}
\examples{
\dontrun{
TODO:
}
}
