% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_hetu_key.R
\name{make_hetu_key}
\alias{make_hetu_key}
\title{Function for creating key table out of personal identification number}
\usage{
make_hetu_key(.data, hetu, first_id = 1, randomize = FALSE)
}
\arguments{
\item{.data}{data which contain henkilotunnus}

\item{hetu}{variable name contain henkilotunnus}

\item{first_id}{first number (integer) which  will starts mapping of henkilotunnus.}

\item{randomize}{if set to TRUE, the order of newly generated ids will be random.}
}
\value{
tibble which contain henkilotunnus mapped to integer.
}
\description{
This functions allows user to create a key table, which maps henkilotunnus to rolling number (id)
The 'henkilotunnus' numbers will be mapped to integers starting from 1 to n.
}
