% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_indicators.R
\name{make_indicators}
\alias{make_indicators}
\title{Filter register based data and add dichotomous variables for each disease.}
\usage{
make_indicators(
  cohort_data,
  reg_data,
  adm_date = NULL,
  disc_date = NULL,
  index_date = NULL,
  time_before = 0,
  time_after = 0,
  idnum,
  codes,
  diag_tbl,
  add_zero_class = TRUE
)
}
\arguments{
\item{cohort_data}{cohort data set containing information about the persons: ID, gender, and index day.}

\item{reg_data}{long format data with variables containing id-numbers, additional date variables, variable holding the ICD, SII or ATC codes, and a variable describing the version of the codes (ICD-9, ICD-10 etc.)}

\item{adm_date}{the variable name of the date of admission}

\item{disc_date}{the name of the variable holding the date of discharge}

\item{index_date}{the name of the variable which is the reference point around which to search}

\item{time_before}{the time before the index date what defines the start of filtering interval. Numeric format, days, e.g, 1 year = 365.25}

\item{time_after}{the time after the index date what defines the start of filtering interval. Numeric format, days, e.g, 1 week = 7}

\item{idnum}{name of the variable holding identification codes (ID) of subjects}

\item{codes}{name of the variable holding ICD, SII, ATC etc. -codes}

\item{diag_tbl}{name of the `regstudies` classification one wants to apply. It can be either `charlson` or `elixhauser`. Alternatively one can use a tibble which holds the classification details}

\item{add_zero_class}{a logical value indicating whether diseases with zero frequencies should be added to the data. Default is TRUE}
}
\value{
returns a tibble object with filtered rows based on a date interval and dichotomous variables for each disease.
}
\description{
Can be used for filtering, e.g., for hospital visits and to add dichotomous variables for each disease.
The returned data set is in wide format.
}
\examples{
\dontrun{
d <- make_indicators(cohort_data = regstudies::sample_cohort, 
                     reg_data = regstudies::sample_regdata,
                     adm_date = adm_date, disc_date = disc_date, index_date = postingdate,
                     time_before = 2*365.25, time_after = 0,
                     idnum = personid, 
                     codes = CODE1, diag_tbl = charlson)

}

}
