% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regstudies_24032020.R
\name{make_regex}
\alias{make_regex}
\title{Make regural expression classifications from LIKE\% classifications}
\usage{
make_regex(.data, classname, diagnosis, diagnosis.rm, sep = ", ")
}
\arguments{
\item{.data}{holds the data which has variable names given as 'classname' and 'diagnosis'. Must be convertible to a tibble.}

\item{classname}{name of the variable containing names/labels for the classes}

\item{diagnosis}{name of the variable containing strings in LIKE\% -format}

\item{diagnosis.rm}{name of the variable defining exceptions of ICD-codes not to be included in the class}

\item{sep}{separator of different diagnoses in 'diagnosis'. String.}
}
\value{
returns a tibble with regular expression definitions
}
\description{
Make regural expression classifications from LIKE\% classifications
}
