% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_class.R
\name{read_class}
\alias{read_class}
\title{Read classes definitions from csv file}
\usage{
read_class(file)
}
\arguments{
\item{file}{the filename or path to the filename as a character string.}
}
\value{
returns classification definintion in long format (tidy data) as that is what the package needs in operation.
}
\description{
Reads classes definitions based on the file path and operates classification definitions from wide format to long format.
}
\examples{
elixhauser_cl <- read_class(file = "data/classification_codes/elixhauser_classes_wide.csv")
}
