% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_classes.R
\name{read_classes}
\alias{read_classes}
\title{Internal function for providing long classification definitions.}
\usage{
read_classes(class_def)
}
\arguments{
\item{classdef}{This variable contains the data in the wide format. See the classifications defined in /datas/classification_codes/}
}
\value{
returns a tibble of classification definition in long format (tidy data) as that is what the package needs in operation.
}
\description{
Operates classification definitions from wide format to long format.
}
\examples{
 \dontrun{
 class_definitions <- vroom(file = "data/elixhauser_classes.csv")
 class_definitions <- read_classes(class_def = class_definitions)
 head(class_definitions)
 }
 
}
\concept{read functions}
