% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_classes_csv.R
\name{read_classes_csv}
\alias{read_classes_csv}
\title{Read class definitions from csv file}
\usage{
read_classes_csv(file)
}
\arguments{
\item{file}{filename or file path as a character string.}
}
\value{
returns classification definintion in long format (tidy data) as that is what the package needs in operation.
}
\description{
Use this function for your own code classification. It reads class definition file based on the file path and operates classification definitions from wide format to long format. You can edit you own classifications in Spreadsheet Program (ex. LibreOffice, Excel) and save file to csv-file.
}
\examples{
 \dontrun{
my_classes <- read_classes_csv(file = "my_classes.csv")
}
}
\seealso{
Other read functions: 
\code{\link{read_classes_tibble}()}
}
\concept{read functions}
