% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_classes_tibble.R
\name{read_classes_tibble}
\alias{read_classes_tibble}
\title{Internal function for providing long classification definitions.}
\usage{
read_classes_tibble(sel_classes)
}
\arguments{
\item{sel_classes}{This variable contains the data in the wide format. See the classifications defined in /datas/classification_codes/}
}
\value{
returns classification definintion in long format (tidy data) as that is what the package needs in operation.
}
\description{
Operates classification definitions from wide format to long format.
}
\examples{
 \dontrun{
 sel_classes <- vroom(file = "data/classification_codes/elixhauser_classes_wide.csv")
 sel_classes2 <- read_classes_tibble(sel_classes = sel_classes)
 head(sel_classes2)
 }
 
}
