% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regdata_pivot_longer.R
\name{regdata_pivot_longer}
\alias{regdata_pivot_longer}
\title{Extends the data with variable describing the code version and returns the data in a long format. This function only works with regular expression tables}
\usage{
regdata_pivot_longer(.data, idnum, diag_tbl = NULL)
}
\arguments{
\item{.data}{tibble in wide-format where class of date variables are dates, class of variables holding atc codes are characters and class of id-numbers is double or int}

\item{idnum}{name of the variable holding identification codes of subjects}

\item{diag_tbl}{tibble which holds the classification details: needs to have variables 'regex' and 'label' 'regex' must hold a string with a regular expression defining classes. 'regex.rm' is optional, defines exceptions to 'regex' (these are not in the group they are named in) 'label' defines the names of the variables of classes (e.g. comorbidity indicators)}
}
\value{
returns a tibble object extended with a variable describing the ICD code version. The data is in long format. ICD codes themself are stored into a new variable \code{CODE}.
}
\description{
The function adds a variable describing the version of the ICD, SII or ATC codes based on the  classification table.
If the ICD, SII or ATC codes are found, the function adds ICD (either icd9 or icd10), ATC or SII code name to the input data set and returns a data in a long format.
This long format data can then be analyzed with functions such as \code{regstudies::classify_elixhauser},
\code{regstudies::classify_charlson} or \code{regstudies::classify_table}.
}
\examples{
\dontrun{
# Use your own icd codes,

my_classes <-
tribble(~class_myclass,         ~label_myclass,                                 ~icd10,                 ~icd9, ~score,
        "aids",          "AIDS or HIV",                  "^B20|^B21|^B22|^B24",      "^042|^043|^044",      7,
        "dementia",             "Dementia", "^F00|^F01|^F02|^F03|^F051|^G30|^G311",    "^290|^2941|^3312",      3,
        "pud", "Peptic ulcer disease",                  "^K25|^K26|^K27|^K28", "^531|^532|^533|^534",      1
)

regdata_long <- regdata \%>\% 
  regdata_pivot_longer(idnum = personid, 
                       diag_tbl = my_classes \%>\% select(icd10, icd9)
  )

D <- regdata_long \%>\% 
  regstudies::classify_codes(codes = CODE, 
                             diag_tbl = read_classes(my_classes)
  ) 
}

}
\author{
Markku Kuismin, Juho Kopra, Jani Miettinen, Reijo Sund
}
