% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stay_length.R
\name{sum_stay_length}
\alias{sum_stay_length}
\title{Count the days in in-hospital care from the start of the washout period.}
\usage{
sum_stay_length(
  .data,
  user_data,
  idnum,
  adm_date,
  disc_date,
  index_date,
  wolen = 365,
  ongoing_end_time = 90
)
}
\arguments{
\item{.data}{cohort data containing at least id-numbers and index days}

\item{user_data}{register data}

\item{idnum}{name of the column/variable holding the person IDs}

\item{adm_date}{the variable name of the date of admission}

\item{disc_date}{the name of the variable holding the date of discharge}

\item{index_date}{the name of the variable which is the reference point around which to search}

\item{wolen}{washout length in days. Default is one year, `365`}

\item{ongoing_end_time}{the time (days) persons were hospitalized at the end of the washout period. Default is three months, `90`}
}
\value{
returns the `.data` tibble extended with two columns: `wo_total_time_hosp` which is the total number of days in hospital during the washout period, and `wo_end_time_hosp` which is the number of days in the hospital at the end of the washout period.
}
\description{
Can be used for filtering hospitalizations during the wash-out period. The function extends the original data `.data` with a column describing the days in-hospital care during the washout-period.
}
\examples{
\dontrun{
Df <- regstudies::sample_regdata \%>\% 
  dplyr::left_join(sample_cohort \%>\% select(personid, postingdate), by="personid")

d <- Df \%>\% regstudies::sum_stay_length(., regstudies::sample_cohort,
                                        idnum = personid,
                                        adm_date = adm_date, 
                                        disc_date = disc_date, 
                                        index_date = postingdate, 
                                        wolen = 2*365, 
                                       ongoing_end_time = 60)
}

}
\keyword{dplyr,}
\keyword{regstudies,}
\keyword{tidyr}
