% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sum_stay_length_v2.R
\name{sum_stay_length_v2}
\alias{sum_stay_length_v2}
\title{Alternative procedure for counting the days in in-hospital care.}
\usage{
sum_stay_length_v2(
  .data,
  user_data,
  idnum,
  adm_date,
  disc_date,
  index_date,
  wolen = 365,
  ongoing_end_time = 90
)
}
\arguments{
\item{.data}{cohort data set containing information about the persons: ID, gender, and index day.}

\item{user_data}{register data}

\item{idnum}{name of the column/variable holding the person IDs}

\item{adm_date}{the variable name of the date of admission}

\item{disc_date}{the name of the variable holding the date of discharge}

\item{index_date}{the name of the variable which is the reference point around which to search}

\item{wolen}{washout length in days. Default is one year, \code{365}}

\item{ongoing_end_time}{the time (days) persons were hospitalized at the end of the washout period. Default is three months, \code{90}}
}
\value{
returns the \code{.data} tibble extended with three columns: \code{wo_total_days} which is the total number of days in hospital during the washout period, \code{wo_end_days} which is the number of days in the hospital at the end of the washout period, and \code{total_days} which is the total number of days in hospital care.
}
\description{
Can be used for filtering hospitalizations during the wash-out period. The function extends the original data \code{.data} with a column describing the days in-hospital care during the washout-period, at the end of it and total number of days the person has been in hospital care.
}
\examples{
\dontrun{
d <- regstudies::sample_cohort \%>\% 
sum_stay_length_v2(., user_data = regstudies::sample_regdata, 
                   idnum = personid, 
                   adm_date = adm_date, 
                   disc_date = disc_date,
                   index_date = postingdate,
                   wolen = 365,
                   ongoing_end_time = 90)
}

}
\author{
Markku Kuismin, Juho Kopra, Jani Miettinen, Reijo Sund
}
