% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_date_in_ival}
\alias{filter_date_in_ival}
\title{Filtering of datetime data within a time range.}
\usage{
filter_date_in_ival(.data, ..., index_date, range = years(2))
}
\arguments{
\item{...}{datetime variables which need to be within the range. May be any number (>1) of variables.  If one variable is within interval, then that row is included in the data (Applies OR operator).}

\item{index_date}{index date (which date variable is to be compared with register data)}

\item{range}{the distance from index date (lubridate format, e.g years(1), weeks(10), days(20) etc.)}
}
\value{
returns a tibble object with filtered rows
}
\description{
Filtering of datetime data within a time range.
}
\examples{
\dontrun{
TODO:
x <- 1
}
}
\seealso{
\code{\link{filter_ival_olap_ival}} for date filtering by different intervals before and after the occurence

Other filter functions: 
\code{\link{filter_date_interval}()},
\code{\link{filter_date_range}()},
\code{\link{filter_ival_olap_ival}()}
}
\concept{filter functions}
