% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/filter.R
\name{filter_ival_olap_ival}
\alias{filter_ival_olap_ival}
\title{Filtering of date interval data within study interval.
Can be used for filtering e.g. for hospital visits, which are date intervals (time periods).
The function finds rows for which \code{ival_var} overlaps with interval created based on \code{index_date}, \code{time_before} and \code{time_after}.}
\usage{
filter_ival_olap_ival(
  .data,
  ival_var,
  index_date,
  time_before = years(2),
  time_after = days(0)
)
}
\arguments{
\item{ival_var}{the interval to be filtered}

\item{index_date}{date or other lubridate variable as reference point around which to search}

\item{time_before}{the time before the index date what defines the start of filtering interval (lubridate format, e.g years(1), weeks(10), days(20) etc.)}

\item{time_after}{the time after the index date (default is 0 days) what defines the start of filtering interval (lubridate format, e.g years(1), weeks(10), days(20) etc.)}
}
\value{
returns a tibble object with filtered rows
}
\description{
Filtering of date interval data within study interval.
Can be used for filtering e.g. for hospital visits, which are date intervals (time periods).
The function finds rows for which \code{ival_var} overlaps with interval created based on \code{index_date}, \code{time_before} and \code{time_after}.
}
\examples{
\dontrun{
d <- dplyr::left_join(sample_cohort,sample_regdata)
filtered_d <- d \%>\% 
  filter_ival_olap_ival(ival_var = lubridate::interval(adm_date,disc_date),
                        index_date = postingdate,
                        time_before = years(2),
                        time_after = days(0))
head(filtered_d)
}
}
\seealso{
\code{\link{filter_date_in_ival}} for date filtering by fixed interval

Other filter functions: 
\code{\link{filter_date_in_ival}()}
}
\concept{filter functions}
