% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_classification_name.R
\name{get_classification_name}
\alias{get_classification_name}
\title{Internal function for getting classification name out of classification tibbles}
\usage{
get_classification_name(.data)
}
\arguments{
\item{.data}{tibble of classification definition}
}
\value{
Returns a string which tells the name of a classification
}
\description{
Internal function for getting classification name out of classification tibbles
}
\seealso{
Other help functions: 
\code{\link{get_na_vars}()},
\code{\link{get_var_types}()},
\code{\link{left_join_replace_na}()},
\code{\link{make_regex}()}
}
\concept{help functions}
