% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable_types.R
\name{get_var_types}
\alias{get_var_types}
\title{Get variable types}
\usage{
get_var_types(.data)
}
\arguments{
\item{.data}{tibble data to be used}
}
\value{
returns a tibble with two character columns: variable and class.
variable tells the variable name and class tells the type of that variable.
}
\description{
This function operates as an tidyverse style version of class() function
}
