% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/left_join_replace_na.R
\name{left_join_replace_na}
\alias{left_join_replace_na}
\title{left_join with fill option}
\usage{
left_join_replace_na(x, y, na_replace_list, ...)
}
\arguments{
\item{x}{tibble data to used as left in a join}

\item{y}{tibble data to used as right in a join}

\item{na_replace_list}{a list definitions of na replacement values for each type of variable (based on 'class' or 'get_var_types' functions)}
}
\value{

}
\description{
Allows user to do left_join and fill missing NA values with 0 or other optional value
}
\examples{
\dontrun{

## Give list of NA values
na_replace_list <- list(
  "logical"=FALSE,
  "character"="",
  "integer"=0L,
  "double"=0.0,
  "numeric"=0,
  "Date"=lubridate::dmy("01-01-1900")
)
testdata \%>\%
  mutate_all(replace_na_by_type,na_replace_list)
}

}
