% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_regex.R
\name{make_regex}
\alias{make_regex}
\title{Make regural expression classifications from LIKE\% classifications}
\usage{
make_regex(.data, classname, diagnosis, diagnosis.rm, sep = ", ")
}
\arguments{
\item{.data}{holds the data which has variable names given as 'classname' and 'diagnosis'. Must be convertible to a tibble.}

\item{classname}{name of the variable containing names/labels for the classes}

\item{diagnosis}{name of the variable containing strings in LIKE\% -format}

\item{diagnosis.rm}{name of the variable defining exceptions of ICD-codes not to be included in the class}

\item{sep}{separator of different diagnoses in 'diagnosis'. String.}
}
\value{
returns a tibble with regular expression definitions
}
\description{
Make regural expression classifications from LIKE\% classifications
}
\examples{
\dontrun{
# We load classification here
excelfile <- readxl::read_excel("./datas/disease_classification.xlsx")
cl9 <- excelfile \%>\%
  make_regex(classname=Lyhenne,diagnosis=sairdiag_icd9,diagnosis.rm=sairdiagpl_icd9) \%>\%
  mutate(icd="icd9")
cl10 <- excelfile \%>\%
  make_regex(classname=Lyhenne,diagnosis=sairdiag,diagnosis.rm=sairdiagpl) \%>\%
  mutate(icd="icd10")

sel_classes<-bind_rows(cl9,cl10) \%>\%
  mutate(classification="own_definition")
}
}
\seealso{
Other help functions: 
\code{\link{get_classification_name}()},
\code{\link{get_na_vars}()},
\code{\link{get_var_types}()},
\code{\link{left_join_replace_na}()},
\code{\link{replace_na_by_type}()}
}
\concept{help functions}
